AABB & collisions detection

+++ src/util/geometry.rs
@@ -1,5 +1,5 @@
 #[derive(Clone)]
 pub enum Shape {
     Circle(f32),
-    Rectangle(i32, i32),
+    Rectangle(f32, f32),
 }
